requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
ca-certificates=2017.08.26=h1d4fec5_0
certifi=2018.1.18=py36_0
cffi=1.11.4=py36h9745a5d_0
cudatoolkit=8.0=3
freetype=2.8=hab7d2ae_1
intel-openmp=2018.0.0=hc7b2577_8
jpeg=9b=h024ee3a_2
libedit=3.1=heed3624_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=h7cc24e2_2
libgfortran-ng=7.2.0=h9f7466a_2
libpng=1.6.34=hb9fc6fc_0
libstdcxx-ng=7.2.0=h7a57d05_2
libtiff=4.0.9=h28f6b97_0
mkl=2018.0.1=h19d6760_4
ncurses=6.0=h9df7e31_2
numpy=1.14.0=py36h3dfced4_1
olefile=0.45.1=py36_0
openssl=1.0.2n=hb7f436b_0
pillow=5.0.0=py36h3deb7b8_0
pip=9.0.1=py36h6c6f9ce_4
pycparser=2.18=py36hf9f622e_1
python=3.6.4=hc3d631a_1
pytorch=0.3.0=py36_cuda8.0.61_cudnn7.0.3h37a80b5_4
readline=7.0=ha6073c6_4
setuptools=38.4.0=py36_0
six=1.11.0=py36h372c433_1
sqlite=3.22.0=h1bed415_0
tk=8.6.7=hc745277_3
torchvision=0.2.0=py36h17b6947_1
wheel=0.30.0=py36hfd4bba0_1
xz=5.2.3=h55aa19d_2
zlib=1.2.11=ha838bed_2
